splitting up ics_test; cleanup
